requirements.txt
Flask
Flask-Login
PyMySQL
Pillow
numpy
torch
h5py
tqdm
opencv-python
python-dotenv